+++ src/main/java/io/github/astrapi69/icon/StringIcon.java
@@ -24,11 +24,20 @@
  */
 package io.github.astrapi69.icon;
 
-import java.awt.*;
+import java.awt.Color;
+import java.awt.Component;
+import java.awt.Font;
+import java.awt.FontMetrics;
+import java.awt.Graphics;
+import java.awt.Graphics2D;
+import java.awt.RenderingHints;
+import java.awt.Toolkit;
 import java.util.Map;
 
-import javax.swing.*;
+import javax.swing.Icon;
+import javax.swing.JComponent;
 
+import io.github.astrapi69.icon.model.IconDimensions;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NonNull;