+++ src/main/java/io/github/astrapi69/icon/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides a bean class that holds information for the icon dimensions
+ */
+package io.github.astrapi69.icon.model;