+++ src/main/java/io/github/astrapi69/icon/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This is the base package for the icon extension and factory classes
+ */
+package io.github.astrapi69.icon;