+++ src/main/java/io/github/astrapi69/img/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This is the package for the image extension class and the enum class for the direction
+ */
+package io.github.astrapi69.img;