+++ src/test/java/io/github/astrapi69/icon/StringIconTest.java
@@ -26,9 +26,10 @@ package io.github.astrapi69.icon;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import javax.swing.*;
+import javax.swing.JLabel;
 
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
 /**
  * Test class for the class {@link StringIcon}