+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 ## Change log
 ----------------------
 
+Version 3
+-------------
+
+ADDED:
+
+- new factory method in class ImageIconFactory that creates a BufferedImage object with custom width, height and a textual description for the image
+- new factory methods for create BufferedImage in class ImageExtensions
+- new package-info classes for javadoc
+
+CHANGED:
+
+- update of jdk version 17
+- update gradle to new version 8.3-rc-4
+- update of gradle plugin dependency io.freefair.gradle:lombok-plugin to new version 8.2.2
+- update of test dependency 'com.github.meanbeanlib:meanbean' to new version 3.0.0-M9
+- code coverage increased
+- moved class IconDimensions to package 'io.github.astrapi69.icon.model'
+
 Version 2.2
 -------------